pyproject.toml
[project]
name = "course-aiagent"
version = "0.1.0"
description = ""
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.52.0",
    "flask>=3.1.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.1.0",
]
